Changes

Return ID's of CQC providers or locations that have reported changes in a given time period.

cqc_changes(
  organisation_type = c("provider", "location"),
  start_date = "2000-01-01",
  end_date = Sys.Date(),
  verbose = TRUE,
  clean_names = TRUE
)

Arguments

organisation_type

One of "provider" or "location".

start_date

The start of the date range, in "YYYY-MM-DD" format, or any vector of character, integer or numeric that can be be parsed with anytime::anytime.

end_date

The end of the date range, in "YYYY-MM-DD" format, or any vector of character, integer or numeric that can be be parsed with anytime::anytime.

verbose

If TRUE prints download progress to console if requesting more than 1,000 records. Defaults to TRUE.

clean_names

If TRUE, converts a column names to snake_case. Defaults to TRUE.

Value

A tibble of all providers or locations that had reported changes in the given period

Examples

# \donttest{
changes_location <- cqc_changes(
  organisation_type = "LoCAtiOn",
  start_date = "2019-03-10",
  end_date = "2019-03-12"
)
# }